services/auth/requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.1
python-dotenv>=1.0.0
gunicorn>=22.0.0
PyJWT>=2.8.0
werkzeug>=3.1.0

